legacy_1 to legacy_2 and legacy to legacy_1
