+++ .github/workflows/_promote_charms_legacy_1.yaml
@@ -59,7 +59,7 @@ jobs:
         uses: canonical/get-workflow-version-action@v1
         with:
           repository-name: canonical/data-platform-workflows
-          file-name: _promote_charms_legacy.yaml
+          file-name: _promote_charms_legacy_1.yaml
           github-token: ${{ secrets.GITHUB_TOKEN }}
       - name: Install CLI
         run: pipx install git+https://github.com/canonical/data-platform-workflows@"${VAR_SHA}"#subdirectory=_cli
@@ -83,7 +83,7 @@ jobs:
           fetch-depth: 0  # Checkout history with git tags
       - name: Promote charms
         id: promote
-        run: promote-charms-legacy --track="${VAR_TRACK}" --from-risk="${VAR_FROM_RISK}" --to-risk="${VAR_TO_RISK}" --ref="${VAR_REF}" --default-branch="${VAR_DEFAULT_BRANCH}"
+        run: promote-charms-legacy-1 --track="${VAR_TRACK}" --from-risk="${VAR_FROM_RISK}" --to-risk="${VAR_TO_RISK}" --ref="${VAR_REF}" --default-branch="${VAR_DEFAULT_BRANCH}"
         env:
           CHARMCRAFT_AUTH: ${{ secrets.charmhub-token }}
           GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
