+++ _cli/data_platform_workflows_cli/craft_tools/promote_legacy_1.py
@@ -50,66 +50,19 @@ class Risk(enum.StrEnum):
         return order.index(self) < order.index(other)
 
 
-def get_commit_sha_and_revisions(
-    *, channel: str, charm_name: str, tag_prefix: str, channel_missing_ok=False
-):
-    """Get (& verify) commit sha that all charm revisions on a Charmhub channel were built from
-
-    Returns (commit sha, charm revisions)
-    """
-    logging.info(f"Getting revisions on {repr(channel)}")
-    response = requests.get(
-        f"https://api.snapcraft.io/v2/charms/info/{charm_name}?fields=channel-map&channel={channel}"
-    )
-    response.raise_for_status()
-    channel_map = response.json()["channel-map"]
-    revisions: list[int] = [item["revision"]["revision"] for item in channel_map]
-    if not revisions:
-        if channel_missing_ok:
-            logging.info(f"No revisions exist on {repr(channel)}")
-            return
-        raise ValueError(f"No revisions exist on {repr(channel)}")
-    logging.info(f"Revisions on {repr(channel)}: {repr(revisions)}")
-
-    logging.info("Checking that revisions were built from the same git commit")
-    commit_shas = set()
-    for revision in revisions:
-        tag = f"{tag_prefix}{revision}"
-        try:
-            commit_shas.add(
-                subprocess.run(
-                    ["git", "rev-list", "-n", "1", tag], capture_output=True, check=True, text=True
-                ).stdout.strip()
-            )
-        except subprocess.CalledProcessError:
-            logging.error(
-                f"Unable to find git tag {repr(tag)}. Was revision {revision} released with "
-                "data-platform-workflows release_charm.yaml?"
-            )
-            raise
-    if len(commit_shas) != 1:
-        raise ValueError(
-            f"Revisions {repr(revisions)} were built from different git commits: "
-            f"{repr(commit_shas)}. Revisions must be built from the same git commit to correctly "
-            "apply git tags for risk (e.g. '14/beta')"
-        )
-    commit_sha = commit_shas.pop()
-    logging.info(f"All revisions on {repr(channel)} were built from git commit {repr(commit_sha)}")
-    return commit_sha, revisions
-
-
-@dataclasses.dataclass(kw_only=True)
-class Metadata:
+@dataclasses.dataclass(frozen=True, kw_only=True)
+class Charm:
+    directory: pathlib.Path
     name: str
     display_name: str
-    oci_resources: dict[str, str]
+    oci_resources: tuple[tuple[str, str]]  # Use instead of dict so that `Charm` is hashable
 
     @classmethod
-    def from_file(cls, *, directory: pathlib.Path):
-        file = yaml.safe_load((directory / "metadata.yaml").read_text())
+    def from_directory(cls, directory: pathlib.Path, /):
+        metadata = yaml.safe_load((directory / "metadata.yaml").read_text())
         # (Only for Kubernetes charms) get OCI resources
         oci_resources = {}
-        for resource_name, resource in file.get("resources", {}).items():
+        for resource_name, resource in metadata.get("resources", {}).items():
             if resource["type"] != "oci-image":
                 continue
             oci_hash = resource["upstream-source"]
@@ -120,11 +73,324 @@ class Metadata:
                 )
             oci_resources[resource_name] = oci_hash
         return cls(
-            name=file["name"], display_name=file["display-name"], oci_resources=oci_resources
+            directory=directory,
+            name=metadata["name"],
+            display_name=metadata["display-name"],
+            oci_resources=tuple(oci_resources.items()),
         )
 
+    @property
+    def tag_prefix(self):
+        if self.directory == pathlib.Path("."):
+            return "rev"
+        else:
+            return f"{self.name}/rev"
 
-def charm():
+
+def get_commit_sha_and_release_title(
+    *, channel: str, charms_: list[Charm], channel_missing_ok=False
+):
+    """Get (& verify) commit sha that all charm revisions on a Charmhub channel name were built from
+
+    Checks revisions across all Charmhub channels (each charm has a Charmhub channel) with name
+
+    Returns (commit sha, GitHub release title)
+    """
+    logging.info(f"Getting revisions on {repr(channel)}")
+    revisions: dict[Charm, list[int]] = {}
+    for charm in charms_:
+        # One-time exception to requirement that charms in monorepo have the same track—to enable
+        # MySQL Router charms to use monorepo. (VM 8.0 track is managed by another team for
+        # historical reasons.) Refresh compatibility tag will still use "dpe".
+        if charm.name == "mysql-router-k8s" and channel.startswith("dpe/"):
+            logging.warning(
+                "Exception for mysql-router-k8s on track 'dpe': using track '8.0' instead for "
+                "Charmhub operations"
+            )
+            charm_channel = channel.replace("dpe/", "8.0/")
+        else:
+            charm_channel = channel
+
+        response = requests.get(
+            f"https://api.snapcraft.io/v2/charms/info/{charm.name}?fields=channel-map&channel={charm_channel}"
+        )
+        response.raise_for_status()
+        channel_map = response.json()["channel-map"]
+        revisions[charm] = sorted(item["revision"]["revision"] for item in channel_map)
+    logging.info(
+        f"Revisions on {repr(channel)}: "
+        f"{repr({charm.name: charm_revisions for charm, charm_revisions in revisions.items()})}"
+    )
+    if channel_missing_ok and all(not charm_revisions for charm_revisions in revisions.values()):
+        logging.info(f"No revisions exist (for any charm) on {repr(channel)}")
+        return None
+    elif missing := next(
+        (charm for charm, charm_revisions in revisions.items() if not charm_revisions), False
+    ):
+        raise ValueError(f"No {repr(missing.name)} revisions exist on {repr(channel)}")
+
+    logging.info("Checking that revisions were built from the same git commit")
+    commit_shas = set()
+    for charm, charm_revisions in revisions.items():
+        for revision in charm_revisions:
+            tag = f"{charm.tag_prefix}{revision}"
+            try:
+                commit_shas.add(
+                    subprocess.run(
+                        ["git", "rev-list", "-n", "1", tag],
+                        capture_output=True,
+                        check=True,
+                        text=True,
+                    ).stdout.strip()
+                )
+            except subprocess.CalledProcessError:
+                logging.error(
+                    f"Unable to find git tag {repr(tag)}. Was {repr(charm.name)} revision "
+                    f"{revision} released with data-platform-workflows release_charm_edge.yaml?"
+                )
+                raise
+    if len(commit_shas) != 1:
+        raise ValueError(
+            "Revisions "
+            f"{repr({charm.name: charm_revisions for charm, charm_revisions in revisions.items()})} "
+            "were built from different git commits: "
+            f"{repr(commit_shas)}. Revisions must be built from the same git commit to correctly "
+            "apply git tags for risk (e.g. '14/beta')"
+        )
+    commit_sha = commit_shas.pop()
+    logging.info(f"All revisions on {repr(channel)} were built from git commit {repr(commit_sha)}")
+
+    # Determine release title
+    if len(charms_) == 1:
+        charm_revisions = revisions[charms_[0]]
+        if len(charm_revisions) == 1:
+            release_title = f"Revision {charm_revisions[0]}"
+        else:
+            release_title = f"Revisions {', '.join(str(revision) for revision in charm_revisions)}"
+    else:
+        charm_names = [charm.name for charm in charms_]
+        assert charm_names == sorted(charm_names)
+        use_substrate_instead_of_name = (
+            len(charm_names) == 2 and charm_names[1] == f"{charm_names[0]}-k8s"
+        )
+        title_parts = []
+        for charm, charm_revisions in revisions.items():
+            if use_substrate_instead_of_name:
+                if charm.name.endswith("-k8s"):
+                    parenthetical = "Kubernetes"
+                else:
+                    parenthetical = "machines"
+            else:
+                parenthetical = charm.name
+            title_parts.append(
+                f"{', '.join(str(revision) for revision in charm_revisions)} ({parenthetical})"
+            )
+        release_title = f"Revisions: {' | '.join(title_parts)}"
+
+    return commit_sha, release_title
+
+
+def get_github_release_tag(*, commit_sha: str) -> str:
+    """Get GitHub release tag from commit"""
+    charm_refresh_compatibility_version_tags = subprocess.run(
+        ["git", "tag", "--list", "v*/*", "--points-at", commit_sha],
+        capture_output=True,
+        check=True,
+        text=True,
+    ).stdout.splitlines()
+    if len(charm_refresh_compatibility_version_tags) != 1:
+        raise ValueError(
+            f"Expected 1 charm refresh compatibility version tags on commit {commit_sha}, got "
+            f"{len(charm_refresh_compatibility_version_tags)} tags: "
+            f"{repr(charm_refresh_compatibility_version_tags)}"
+        )
+    return charm_refresh_compatibility_version_tags[0]
+
+
+def get_legacy_release_tag_for_commit(*, commit_sha: str, charms_: list[Charm]) -> str:
+    """Get alphabetically-first revision tag on commit as the GitHub release tag.
+
+    Used when refresh_versions.toml does not exist (no v*/* tags).
+    """
+    all_tags = []
+    for charm in charms_:
+        tags = subprocess.run(
+            ["git", "tag", "--list", f"{charm.tag_prefix}*", "--points-at", commit_sha],
+            capture_output=True,
+            check=True,
+            text=True,
+        ).stdout.splitlines()
+        all_tags.extend(tags)
+    if not all_tags:
+        raise ValueError(
+            f"No revision tags found on commit {commit_sha} for charms "
+            f"{[c.name for c in charms_]}"
+        )
+    # Pick alphabetically first tag because of
+    # https://github.com/orgs/community/discussions/149281#discussioncomment-12071170
+    return sorted(all_tags)[0]
+
+
+def get_last_stable_release_tag(*, track: str, charms_: list[Charm]) -> str | None:
+    """Get GitHub release tag for last stable release (if it exists) on `track`"""
+    channel = f"{track}/{Risk.STABLE.value}"
+    logging.info(f"Getting GitHub release tag for last {repr(channel)} release")
+    output = get_commit_sha_and_release_title(
+        channel=channel,
+        charms_=charms_,
+        channel_missing_ok=True,  # In case no previous stable release
+    )
+    if output is None:
+        logging.warning(f"No existing release found on {repr(channel)}")
+        return None
+    commit_sha, _ = output
+    tag = get_github_release_tag(commit_sha=commit_sha)
+    logging.info(f"GitHub release tag for last {repr(channel)} release: {repr(tag)}")
+    return tag
+
+
+def get_legacy_last_stable_release_tag(*, track: str, charms_: list[Charm]) -> str | None:
+    """Get GitHub release tag for last stable release using revision tags.
+
+    Used when refresh_versions.toml does not exist (no v*/* tags).
+    """
+    channel = f"{track}/{Risk.STABLE.value}"
+    logging.info(f"Getting GitHub release tag for last {repr(channel)} release")
+    output = get_commit_sha_and_release_title(
+        channel=channel,
+        charms_=charms_,
+        channel_missing_ok=True,  # In case no previous stable release
+    )
+    if output is None:
+        logging.warning(f"No existing release found on {repr(channel)}")
+        return None
+    commit_sha, _ = output
+    tag = get_legacy_release_tag_for_commit(commit_sha=commit_sha, charms_=charms_)
+    logging.info(f"GitHub release tag for last {repr(channel)} release: {repr(tag)}")
+    return tag
+
+
+def oxford_comma(items: list[str], /):
+    match items:
+        case []:
+            raise ValueError("List must not be empty")
+        case [first]:
+            return first
+        case [first, second]:
+            return f"{first} and {second}"
+        case [*all_but_last, last]:
+            return f"{', '.join(all_but_last)}, and {last}"
+
+
+def _validate_promotion_and_create_release(
+    *,
+    dry_run: bool,
+    charms_: list[Charm],
+    track: str,
+    channel: str,
+    to_risk: Risk,
+    ref: str,
+    default_branch: str,
+):
+    if dry_run:
+        logging.info("Checking that revisions that will be promoted are from the same git commit")
+    else:
+        logging.info("Getting git commit of revisions that were promoted")
+    promoted_commit_sha, release_title = get_commit_sha_and_release_title(
+        channel=channel, charms_=charms_
+    )
+
+    subprocess.run(["git", "checkout", promoted_commit_sha], check=True)
+
+    for charm in charms_:
+        try:
+            # Check that OCI images are pinned to sha256 digest
+            promoted_charm = Charm.from_directory(charm.directory)
+        except FileNotFoundError:
+            if dry_run:
+                message = (
+                    f"Charm at {repr(charm.directory)} exists on latest commit on branch but does "
+                    f"not exist on commit on {repr(channel)}"
+                )
+            else:
+                message = (
+                    f"Charm at {repr(charm.directory)} was deleted or moved during promotion. "
+                    "Invalid charm promoted"
+                )
+            raise FileNotFoundError(message)
+
+        if promoted_charm.name != charm.name:
+            if dry_run:
+                message = (
+                    f"Charm 'name' in metadata.yaml changed between latest commit on branch "
+                    f"({repr(charm.name)}) and commit on {repr(channel)} "
+                    f"({repr(promoted_charm.name)}). Unable to promote charm"
+                )
+            else:
+                message = (
+                    "Charm 'name' in metadata.yaml changed while charm was promoted. Invalid "
+                    f"charm promoted. Expected charm name {repr(charm.name)}, got "
+                    f"{repr(promoted_charm.name)} instead"
+                )
+            raise ValueError(message)
+
+    github_release_tag = get_github_release_tag(commit_sha=promoted_commit_sha)
+
+    if to_risk is Risk.CANDIDATE:
+        if dry_run:
+            logging.info(
+                "Checking that the last stable release revisions are from the same git commit and "
+                "that we can determine the GitHub release tag"
+            )
+        previous_github_release_tag = get_last_stable_release_tag(track=track, charms_=charms_)
+
+    if dry_run:
+        return
+
+    if to_risk is Risk.CANDIDATE:
+        charm_display_names = oxford_comma(
+            [f"[{charm.display_name}](https://charmhub.io/{charm.name})" for charm in charms_]
+        )
+        # Say "/stable" in release notes so that it's correct when the release is published
+        prepended_notes = f"""A new revision of {charm_display_names} has been published in the {track}/{Risk.STABLE.value} channel on Charmhub.
+"""
+        for charm in charms_:
+            if not charm.oci_resources:
+                continue
+            prepended_notes += f"\n{charm.name} OCI image resources:\n"
+            for resource_name, source in charm.oci_resources:
+                prepended_notes += f"- `{resource_name}={source}`\n"
+        command = [
+            "gh",
+            "release",
+            "create",
+            github_release_tag,
+            "--verify-tag",
+            "--draft",
+            "--generate-notes",
+            "--title",
+            release_title,
+            "--notes",
+            prepended_notes,
+        ]
+        if previous_github_release_tag is not None:
+            command.extend(("--notes-start-tag", previous_github_release_tag))
+        logging.info("Creating GitHub draft release")
+        subprocess.run(command, check=True)
+    elif to_risk is Risk.STABLE:
+        # Publish GitHub release draft created during promotion to candidate risk
+        logging.info("Publishing GitHub release")
+        command = ["gh", "release", "edit", github_release_tag, "--verify-tag", "--draft=false"]
+        # Check if the branch this workflow run was triggered on is the default branch for this
+        # GitHub repository
+        if ref == f"refs/heads/{default_branch}":
+            command.append("--latest")
+        else:
+            command.append("--latest=false")
+        subprocess.run(command, check=True)
+
+
+def charms():
     parser = argparse.ArgumentParser()
     parser.add_argument("--track", required=True)
     parser.add_argument("--from-risk", required=True)
@@ -132,15 +398,19 @@ def charm():
     parser.add_argument("--ref", required=True)
     parser.add_argument("--default-branch", required=True)
     args = parser.parse_args()
-    directory = pathlib.Path(".")
 
-    # Check if a refresh_versions.toml file exists anywhere in the repository
-    if next(pathlib.Path().glob("**/refresh_versions.toml"), False):
+    charms_: list[Charm] = []
+    for path in pathlib.Path().glob("**/charmcraft.yaml"):
+        if "tests" in path.parts:
+            logging.info(f"Ignoring charm inside a 'tests' directory: {repr(path.parent)}")
+            continue
+        charms_.append(Charm.from_directory(path.parent))
+
+    if len({charm.name for charm in charms_}) != len(charms_):
         raise ValueError(
-            "The `_promote_charm_legacy_1.yaml` workflow does not support tracks with charm refresh "
-            "compatibility version tags. Use `_promote_charms.yaml` instead: "
-            "https://github.com/canonical/data-platform-workflows/blob/main/.github/workflows/_promote_charms.md"
+            f"Duplicate charms with same 'name' in metadata.yaml not supported: {repr(charms_)}"
         )
+    charms_ = sorted(charms_, key=lambda charm: charm.name)
 
     track = args.track
     if track == "":
@@ -172,158 +442,63 @@ def charm():
         raise FileNotFoundError(
             "Repository must contain `.github/release.yaml` to automatically generate release "
             "notes in the correct format. See "
-            "https://github.com/canonical/data-platform-workflows/blob/main/.github/workflows/_promote_charm_legacy_1.md#step-3-add-githubreleaseyaml-file"
+            "https://github.com/canonical/data-platform-workflows/blob/main/.github/workflows/_promote_charms_legacy_1.md#step-3-add-githubreleaseyaml-file"
         )
 
     from_channel = f"{track}/{from_risk}"
     to_channel = f"{track}/{to_risk}"
 
-    charm_name = yaml.safe_load((directory / "metadata.yaml").read_text())["name"]
-    # `tag_prefix` format from release_charm.yaml
-    if directory == pathlib.Path("."):
-        tag_prefix = "rev"
-    else:
-        tag_prefix = f"{charm_name}/rev"
-
-    logging.info("Checking that revisions that will be promoted are from the same git commit")
-    # Check before promoting so that we fail early if `from_channel` revisions were built from
-    # different git commits.
-    from_commit_sha, _ = get_commit_sha_and_revisions(
-        channel=from_channel, charm_name=charm_name, tag_prefix=tag_prefix
-    )
-    if to_risk is Risk.CANDIDATE:
-        # We will need to get commit for most recent stable release
-        stable_channel = f"{track}/{Risk.STABLE.value}"
-        logging.info(f"Checking that {repr(stable_channel)} revisions are from the same git commit")
-        get_commit_sha_and_revisions(
-            channel=stable_channel,
-            charm_name=charm_name,
-            tag_prefix=tag_prefix,
-            channel_missing_ok=True,  # In case no previous stable release
-        )
-
-    subprocess.run(["git", "checkout", from_commit_sha], check=True)
-    # Don't reuse commit sha to avoid race condition if `from_channel` changes before
-    # `charmcraft promote` is run. Instead, get commit sha again from `to_channel` after promoting
-    del from_commit_sha
-
-    # Check that OCI images are pinned to sha256 digest
-    metadata_on_from_commit = Metadata.from_file(directory=directory)
-
-    if metadata_on_from_commit.name != charm_name:
-        raise ValueError(
-            "Charm name in metadata.yaml changed between latest commit on branch "
-            f"({repr(charm_name)}) and commit on {repr(from_channel)} "
-            f"({repr(metadata_on_from_commit.name)}). Unable to promote charm"
-        )
-    # Don't reuse OCI image digests to avoid race condition if `from_channel` changes before
-    # `charmcraft promote` is run. Instead, get OCI image digests again from `to_channel` after
-    # promoting
-    del metadata_on_from_commit
-
-    logging.info(
-        f"Promoting {repr(charm_name)} charm from {repr(from_channel)} to {repr(to_channel)}"
-    )
-    subprocess.run(
-        [
-            "charmcraft",
-            "promote",
-            "--name",
-            charm_name,
-            "--from-channel",
-            from_channel,
-            "--to-channel",
-            to_channel,
-            "--yes",
-        ],
-        check=True,
+    _validate_promotion_and_create_release(
+        dry_run=True,
+        charms_=charms_,
+        track=track,
+        channel=from_channel,
+        to_risk=to_risk,
+        ref=ref,
+        default_branch=default_branch,
     )
 
-    logging.info("Getting git commit of revisions that were promoted")
-    promoted_commit_sha, charm_revisions = get_commit_sha_and_revisions(
-        channel=to_channel, charm_name=charm_name, tag_prefix=tag_prefix
-    )
-
-    subprocess.run(["git", "checkout", promoted_commit_sha], check=True)
-
-    metadata = Metadata.from_file(directory=directory)
-    if metadata.name != charm_name:
-        raise ValueError(
-            "Charm name in metadata.yaml changed while charm was promoted. Invalid charm "
-            f"promoted. Expected charm name {repr(charm_name)}, got {repr(metadata.name)} instead"
-        )
-
-    possible_release_tags = [f"{tag_prefix}{revision}" for revision in charm_revisions]
-    # Pick alphabetically first tag because of
-    # https://github.com/orgs/community/discussions/149281#discussioncomment-12071170
-    github_release_tag = sorted(possible_release_tags)[0]
-
-    if to_risk is Risk.CANDIDATE:
-        # Get tag for most recent stable release
-        stable_channel = f"{track}/{Risk.STABLE.value}"
-        logging.info(f"Getting git tag for most recent {repr(stable_channel)} release")
-        output = get_commit_sha_and_revisions(
-            channel=stable_channel,
-            charm_name=charm_name,
-            tag_prefix=tag_prefix,
-            channel_missing_ok=True,  # In case no previous stable release
-        )
-        if output is None:
-            logging.warning(f"No existing release found on {repr(stable_channel)}")
-            previous_github_release_tag = None
-        else:
-            _, stable_revisions = output
-            possible_stable_release_tags = [
-                f"{tag_prefix}{revision}" for revision in stable_revisions
-            ]
-            # Pick alphabetically first tag because of
-            # https://github.com/orgs/community/discussions/149281#discussioncomment-12071170
-            previous_github_release_tag = sorted(possible_stable_release_tags)[0]
-            logging.info(
-                f"Found git tag for most recent {repr(stable_channel)} release: "
-                f"{repr(previous_github_release_tag)}"
+    logging.info(f"Promoting charms from {repr(from_channel)} to {repr(to_channel)}")
+    for charm in charms_:
+        # One-time exception to requirement that charms in monorepo have the same track—to enable
+        # MySQL Router charms to use monorepo. (VM 8.0 track is managed by another team for
+        # historical reasons.) Refresh compatibility tag will still use "dpe".
+        if charm.name == "mysql-router-k8s" and track == "dpe":
+            logging.warning(
+                "Exception for mysql-router-k8s on track 'dpe': using track '8.0' instead for "
+                "Charmhub operations"
             )
+            charm_from_channel = from_channel.replace("dpe/", "8.0/")
+            charm_to_channel = to_channel.replace("dpe/", "8.0/")
+        else:
+            charm_from_channel = from_channel
+            charm_to_channel = to_channel
 
-        if len(charm_revisions) > 1:
-            title = f"Revisions {', '.join(str(revision) for revision in sorted(charm_revisions))}"
-        else:
-            title = f"Revision {charm_revisions[0]}"
-        # Say "/stable" in release notes so that it's correct when the release is published
-        prepended_notes = f"""A new revision of {metadata.display_name} has been published in the {track}/{Risk.STABLE.value} channel on [Charmhub](https://charmhub.io/{charm_name}).
-"""
-        if metadata.oci_resources:
-            prepended_notes += "\nOCI image resources:\n"
-            for resource_name, source in metadata.oci_resources.items():
-                prepended_notes += f"- `{resource_name}={source}`\n"
-        command = [
-            "gh",
-            "release",
-            "create",
-            github_release_tag,
-            "--verify-tag",
-            "--draft",
-            "--generate-notes",
-            # TODO: remove when charm refresh versioning implemented?
-            # If `--draft` is passed, `--latest` appears to have no affect on release editing
-            # in GitHub's UI (the UI will set latest by default regardless)
-            "--latest=false",
-            "--title",
-            title,
-            "--notes",
-            prepended_notes,
-        ]
-        if previous_github_release_tag is not None:
-            command.extend(("--notes-start-tag", previous_github_release_tag))
-        logging.info("Creating GitHub draft release")
-        subprocess.run(command, check=True)
-    elif to_risk is Risk.STABLE:
-        # Publish GitHub release draft created during promotion to candidate risk
-        logging.info("Publishing GitHub release")
-        command = ["gh", "release", "edit", github_release_tag, "--verify-tag", "--draft=false"]
-        # Check if the branch this workflow run was triggered on is the default branch for this
-        # GitHub repository
-        if ref == f"refs/heads/{default_branch}":
-            command.append("--latest")
-        else:
-            command.append("--latest=false")
-        subprocess.run(command, check=True)
+        logging.info(
+            f"Promoting {repr(charm.name)} charm from {repr(charm_from_channel)} to "
+            f"{repr(charm_to_channel)}"
+        )
+        subprocess.run(
+            [
+                "charmcraft",
+                "promote",
+                "--name",
+                charm.name,
+                "--from-channel",
+                charm_from_channel,
+                "--to-channel",
+                charm_to_channel,
+                "--yes",
+            ],
+            check=True,
+        )
+
+    _validate_promotion_and_create_release(
+        dry_run=False,
+        charms_=charms_,
+        track=track,
+        channel=to_channel,
+        to_risk=to_risk,
+        ref=ref,
+        default_branch=default_branch,
+    )
