+++ _cli/pyproject.toml
@@ -27,8 +27,8 @@ parse-snap-version = "data_platform_workflows_cli.parse_snap_version:main"
 sync-docs = "data_platform_workflows_cli.sync_docs:main"
 compute-path-in-artifact = "data_platform_workflows_cli.compute_path_in_artifact:main"
 promote-charms = "data_platform_workflows_cli.craft_tools.promote:charms"
-promote-charms-legacy = "data_platform_workflows_cli.craft_tools.promote_legacy:charms"
-promote-charm-legacy-1 = "data_platform_workflows_cli.craft_tools.promote_legacy_1:charm"
+promote-charms-legacy-1 = "data_platform_workflows_cli.craft_tools.promote_legacy_1:charms"
+promote-charm-legacy-2 = "data_platform_workflows_cli.craft_tools.promote_legacy_2:charm"
 check-python-package-pr-title = "data_platform_workflows_cli.check_semantic_version_prefix:check_pr_title"
 create-semantic-version-tag = "data_platform_workflows_cli.create_semantic_version_tag:main"
 create-charm-version-tag-edge = "data_platform_workflows_cli.create_charm_refresh_version_tag_edge:main"
